Changes -
I added a scanner to iterate over the string in the text box
and separate it by line. I'm not sure what to do from here so I'm
handing it over to you guys. My main problem is that i'm not sure
how to deal with classes that have no meeting time or week days listed.